requirements.txt
python-telegram-bot
openai
langchain
tiktoken
